+++ pkg/types/testutils/chain_registry_faker.go
@@ -7,12 +7,13 @@ import (
 	"github.com/consensys/orchestrate/pkg/toolkit/app/multitenancy"
 	"github.com/consensys/orchestrate/pkg/types/api"
 	"github.com/consensys/orchestrate/pkg/types/entities"
+	"github.com/consensys/quorum-key-manager/pkg/common"
 	"github.com/gofrs/uuid"
 )
 
 func FakeRegisterChainRequest() *api.RegisterChainRequest {
 	return &api.RegisterChainRequest{
-		Name: "mainnet",
+		Name: "chain-" + common.RandString(5),
 		URLs: []string{"http://chain:8545"},
 		Listener: api.RegisterListenerRequest{
 			FromBlock:         "latest",
@@ -23,20 +24,20 @@ func FakeRegisterChainRequest() *api.RegisterChainRequest {
 			Type: entities.EEAChainType,
 		},
 		Labels: map[string]string{
-			"label1": "val1",
-			"label2": "val2",
+			"label1": common.RandString(5),
+			"label2": common.RandString(5),
 		},
 	}
 }
 
 func FakeUpdateChainRequest() *api.UpdateChainRequest {
 	return &api.UpdateChainRequest{
-		Name: "mainnet",
+		Name: "chain-" + common.RandString(5),
 		Listener: &api.UpdateListenerRequest{
 			CurrentBlock: 55,
 		},
 		Labels: map[string]string{
-			"label3": "val3",
+			"label3": common.RandString(5),
 		},
 	}
 }
@@ -44,7 +45,7 @@ func FakeUpdateChainRequest() *api.UpdateChainRequest {
 func FakeChainResponse() *api.ChainResponse {
 	return &api.ChainResponse{
 		UUID:                      uuid.Must(uuid.NewV4()).String(),
-		Name:                      "ganache",
+		Name:                      "chain-" + common.RandString(5),
 		TenantID:                  multitenancy.DefaultTenant,
 		URLs:                      []string{"http://ethereum-node:8545"},
 		ChainID:                   big.NewInt(888),