+++ services/api/business/use-cases/faucets/update_faucet.go
@@ -10,6 +10,7 @@ import (
 	"github.com/consensys/orchestrate/services/api/business/parsers"
 	usecases "github.com/consensys/orchestrate/services/api/business/use-cases"
 	"github.com/consensys/orchestrate/services/api/store"
+	ethcommon "github.com/ethereum/go-ethereum/common"
 )
 
 const updateFaucetComponent = "use-cases.update-faucet"
@@ -33,6 +34,24 @@ func (uc *updateFaucetUseCase) Execute(ctx context.Context, faucet *entities.Fau
 	ctx = log.WithFields(ctx, log.Field("faucet", faucet.UUID))
 	logger := uc.logger.WithContext(ctx)
 	logger.Debug("updating faucet")
+
+	if faucet.ChainRule != "" {
+		_, err := uc.db.Chain().FindOneByUUID(ctx, faucet.ChainRule, userInfo.AllowedTenants, userInfo.Username)
+		if errors.IsNotFoundError(err) {
+			return nil, errors.InvalidParameterError("cannot find new linked chain")
+		} else if err != nil {
+			return nil, err
+		}
+	}
+
+	if faucet.CreditorAccount.String() != new(ethcommon.Address).String() {
+		_, err := uc.db.Account().FindOneByAddress(ctx, faucet.CreditorAccount.String(), userInfo.AllowedTenants, userInfo.Username)
+		if errors.IsNotFoundError(err) {
+			return nil, errors.InvalidParameterError("cannot find updated creditor account")
+		} else if err != nil {
+			return nil, err
+		}
+	}
 
 	faucetModel := parsers.NewFaucetModelFromEntity(faucet)
 	faucetModel.TenantID = userInfo.TenantID