+++ services/api/business/use-cases/faucets/update_faucet_test.go
@@ -7,9 +7,11 @@ import (
 	"testing"
 
 	"github.com/consensys/orchestrate/pkg/toolkit/app/multitenancy"
+	"github.com/consensys/orchestrate/services/api/business/parsers"
+	"github.com/consensys/orchestrate/services/api/store/models/testutils"
+	"github.com/stretchr/testify/require"
 
 	testutils2 "github.com/consensys/orchestrate/pkg/types/testutils"
-	"github.com/consensys/orchestrate/services/api/business/parsers"
 	"github.com/consensys/orchestrate/services/api/store/mocks"
 
 	"github.com/consensys/orchestrate/pkg/errors"
@@ -23,48 +25,72 @@ func TestUpdateFaucet_Execute(t *testing.T) {
 	defer ctrl.Finish()
 
 	mockDB := mocks.NewMockDB(ctrl)
+	accountAgent := mocks.NewMockAccountAgent(ctrl)
+	chainAgent := mocks.NewMockChainAgent(ctrl)
 	faucetAgent := mocks.NewMockFaucetAgent(ctrl)
-	mockDB.EXPECT().Faucet().Return(faucetAgent).AnyTimes()
-	faucet := testutils2.FakeFaucet()
+	mockDB.EXPECT().Chain().AnyTimes().Return(chainAgent)
+	mockDB.EXPECT().Account().AnyTimes().Return(accountAgent)
+	mockDB.EXPECT().Faucet().AnyTimes().Return(faucetAgent)
+
+	chain := testutils.FakeChainModel()
+	account := testutils.FakeAccountModel()
 	userInfo := multitenancy.NewUserInfo("tenantOne", "username")
 	usecase := NewUpdateFaucetUseCase(mockDB)
+	faucet := testutils2.FakeFaucet()
+	faucet.TenantID = userInfo.TenantID
+	faucetModel := parsers.NewFaucetModelFromEntity(faucet)
 
+	expectedErr := errors.NotFoundError("error")
 	t.Run("should execute use case successfully", func(t *testing.T) {
-		faucetModel := parsers.NewFaucetModelFromEntity(faucet)
-		faucetModel.TenantID = userInfo.TenantID
+
+		chainAgent.EXPECT().FindOneByUUID(gomock.Any(), faucet.ChainRule, userInfo.AllowedTenants, userInfo.Username).
+			Return(chain, nil)
+		accountAgent.EXPECT().FindOneByAddress(gomock.Any(), faucetModel.CreditorAccount, userInfo.AllowedTenants,
+			userInfo.Username).Return(account, nil)
 		faucetAgent.EXPECT().Update(gomock.Any(), faucetModel, userInfo.AllowedTenants).Return(nil)
 		faucetAgent.EXPECT().FindOneByUUID(gomock.Any(), faucet.UUID, userInfo.AllowedTenants).Return(faucetModel, nil)
 
 		resp, err := usecase.Execute(ctx, faucet, userInfo)
 
 		assert.NoError(t, err)
-		assert.Equal(t, parsers.NewFaucetFromModel(faucetModel), resp)
+		assert.Equal(t, faucet, resp)
 	})
 
-	t.Run("should fail with same error if update faucet fails", func(t *testing.T) {
-		expectedErr := errors.NotFoundError("error")
-
-		faucetAgent.EXPECT().Update(gomock.Any(), gomock.Any(), userInfo.AllowedTenants).Return(expectedErr)
+	t.Run("should fail with the same error if update faucet fails", func(t *testing.T) {
+		chainAgent.EXPECT().FindOneByUUID(gomock.Any(), faucetModel.ChainRule, userInfo.AllowedTenants, userInfo.Username).
+			Return(chain, nil)
+		accountAgent.EXPECT().FindOneByAddress(gomock.Any(), faucetModel.CreditorAccount, userInfo.AllowedTenants,
+			userInfo.Username).Return(account, nil)
+		faucetAgent.EXPECT().Update(gomock.Any(), faucetModel, userInfo.AllowedTenants).Return(expectedErr)
 
 		resp, err := usecase.Execute(ctx, faucet, userInfo)
 
 		assert.Nil(t, resp)
-		assert.Error(t, err)
-		assert.Equal(t, errors.FromError(expectedErr).ExtendComponent(updateFaucetComponent), err)
+		require.Error(t, err)
+		assert.Equal(t, expectedErr, err)
 	})
 
-	t.Run("should fail with same error if findOne faucet fails", func(t *testing.T) {
-		faucetModel := parsers.NewFaucetModelFromEntity(faucet)
-		faucetModel.TenantID = userInfo.TenantID
-		expectedErr := errors.NotFoundError("error")
-
-		faucetAgent.EXPECT().Update(gomock.Any(), faucetModel, userInfo.AllowedTenants).Return(nil)
-		faucetAgent.EXPECT().FindOneByUUID(gomock.Any(), faucet.UUID, userInfo.AllowedTenants).Return(nil, expectedErr)
+	t.Run("should fail with invalid params error if cannot find account", func(t *testing.T) {
+		chainAgent.EXPECT().FindOneByUUID(gomock.Any(), faucet.ChainRule, userInfo.AllowedTenants, userInfo.Username).
+			Return(chain, nil)
+		accountAgent.EXPECT().FindOneByAddress(gomock.Any(), faucet.CreditorAccount.String(), userInfo.AllowedTenants,
+			userInfo.Username).Return(nil, expectedErr)
 
 		resp, err := usecase.Execute(ctx, faucet, userInfo)
 
 		assert.Nil(t, resp)
-		assert.Error(t, err)
-		assert.Equal(t, errors.FromError(expectedErr).ExtendComponent(updateFaucetComponent), err)
+		require.Error(t, err)
+		assert.True(t, errors.IsInvalidParameterError(err))
+	})
+
+	t.Run("should fail with invalid params error if cannot find chain", func(t *testing.T) {
+		chainAgent.EXPECT().FindOneByUUID(gomock.Any(), faucet.ChainRule, userInfo.AllowedTenants, userInfo.Username).
+			Return(nil, expectedErr)
+
+		resp, err := usecase.Execute(ctx, faucet, userInfo)
+
+		assert.Nil(t, resp)
+		require.Error(t, err)
+		assert.True(t, errors.IsInvalidParameterError(err))
 	})
 }